refactor: Refactor state to nanosotores
